% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCODE.R
\name{cv_lambda}
\alias{cv_lambda}
\title{Cross-validation for sparsity parameter lambda}
\usage{
cv_lambda(data, time, ode.model, par.names, state.names, \cr        par.initial, basis.list,lambda,lambda_grid,cv_points,controls)
}
\arguments{
\item{data}{A data frame or matrix contrain observations from each dimension of the ODE model.}

\item{time}{The vector contain observation times or a matrix if time points are different between dimensions.}

\item{ode.model}{Defined R function that computes the time derivative of the ODE model given observations of states variable.}

\item{par.names}{The names of structural parameters defined in the 'ode.model'.}

\item{state.names}{The names of state variables defined in the 'ode.model'.}

\item{par.initial}{Initial value of structural parameters to be optimized.}

\item{basis.list}{A list of basis objects for smoothing each dimension's observations. Can be the same or different across dimensions.}

\item{lambda_grid}{A search grid for finding the optimial sparsity parameter lambda.}

\item{cv_points}{A number indicating how many data points (>= 5) will be saved for doing cross validation. Default is set at 5 as minimum.}

\item{controls}{A list of control parameters. See ‘Details’.}
}
\value{
\item{lambda_grid}{The original input vector of a search grid for the optimal lambda.}

\item{cv.score}{The vector contains the cross validation score for each lambda}
}
\description{
Obtain the optimal sparsity parameter given a search grid based on cross validation score.
}
