\name{xmisc-package}
\alias{xmisc-package}
\alias{xmisc}
\docType{package}
\title{
  Xing's Miscellaneous Functions
}
\description{
  This package includes some utility functions used by Xing Qiu.
}
\details{
\tabular{ll}{
Package: \tab xmisc\cr
Type: \tab Package\cr
Version: \tab 0.01\cr
Date: \tab 2009-04-27\cr
License: \tab GPL 2.0\cr
LazyLoad: \tab yes\cr
}
Please see the help files of individual functions.
}
\author{
  Xing Qiu
Maintainer: Xing Qiu <xqiu@bst.rochester.edu>
}
\references{
  Some text book.
}
\keyword{package}
\examples{
% X <- randomization(50, ngroups=2, nblocks=5)
% randomization.latex(X)
}
