% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pCODE.R
\name{innerobj_lkh}
\alias{innerobj_lkh}
\title{Inner objective function (likelihood and multiple dimension version)}
\usage{
innerobj_lkh(basis_coef, ode.par, input, derive.model, likelihood.fun)
}
\arguments{
\item{basis_coef}{Basis coefficients for interpolating observations given a basis boject.}

\item{ode.par}{Structural parameters of the ODD model.}

\item{input}{Contains dependencies for the optimization, including observations, ode penalty, and etc..}

\item{derive.model}{The function defines the ODE model and is the same as the ode.model in 'pcode'.}

\item{likelihood.fun}{The likelihood or loglikelihood function of the errors.}
}
\value{
obj.eval The evaluation of the inner objective function.
}
\description{
An objective function combines the likelihood or loglikelihood of errors from each dimension of state variables and the penalty controls how the state estimates fail to satisfy the ODE model.
}
