% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCODE.R
\name{innerobj_multi}
\alias{innerobj_multi}
\title{Inner objective function (multiple dimension version)}
\usage{
innerobj_multi(basis_coef, ode.par, input, derive.model,NLS)
}
\arguments{
\item{basis_coef}{Basis coefficients for interpolating observations given a basis object.}

\item{ode.par}{Structural parameters of the ODE model.}

\item{input}{Contains dependencies for the optimization, including observations, penalty parameter lambda, and etc..}

\item{derive.model}{The function defines the ODE model and is the same as the \code{ode.model} in \code{pcode}.}

\item{NLS}{Default is \code{TRUE} so the function returns vector of residuals, and otherwise returns sum of squared errors.}
}
\value{
\item{residual.vec}{Vector of residuals and evaluation of penalty function on quadrature points for approximating the integral.}
}
\description{
An objective function combines the sum of squared error of basis expansion estimates and the penalty controls how those estimates fail to satisfies the ODE model
}
