% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCODE.R
\name{outterobj_lkh}
\alias{outterobj_lkh}
\title{Outter objective function (likelihood and multiple dimension version)}
\usage{
outterobj_lkh(ode.parameter, basis.initial, derivative.model, likelihood.fun, inner.input)
}
\arguments{
\item{ode.parameter}{Structural parameters of the ODE model.}

\item{basis.initial}{Initial values of the basis coefficients for nonlinear least square optimization.}

\item{derivative.model}{The function defines the ODE model and is the same as the ode.model in 'pcode'}

\item{likelihood.fun}{The likelihood or loglikelihood function of the errors.}

\item{inner.input}{Input that will be passed to the inner objective function. Contains dependencies for the optimization, including observations, penalty parameter lambda, and etc..}
}
\value{
neglik The negative of the likelihood or the loglikelihood function that will be passed further to the 'optim' function.
}
\description{
An objective function of the structural parameter computes the measure of fit.
}
