% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PCODE.R
\name{PC_ODE}
\alias{PC_ODE}
\title{Parameter Cascade Method for Ordinary Differential Equation Models}
\usage{
PC_ODE(data, time, ode.model, par.names, state.names, \cr        par.initial, basis.list,lambda,controls)
}
\arguments{
\item{data}{A data frame or a matrix contain observations from each dimension of the ODE model.}

\item{time}{The vector contain observation times or a matrix if time points are different between dimensions.}

\item{ode.model}{Defined R function that computes the time derivative of the ODE model given observations of states variable.}

\item{par.names}{The names of structural parameters defined in the 'ode.model'.}

\item{state.names}{The names of state variables defined in the 'ode.model'.}

\item{par.initial}{Initial value of structural parameters to be optimized.}

\item{basis.list}{A list of basis objects for smoothing each dimension's observations. Can be the same or different across dimensions.}

\item{lambda}{Penalty parameter.}

\item{controls}{A list of control parameters. See ‘Details’.}
}
\value{
\item{structural.par}{The structural parameters of the ODE model.}

\item{nuissance.par}{The nuissance parameters or the basis coefficients for interpolating observations.}
}
\description{
Obtain estiamtes of structural parameters of an ODE model by parameter cascade method.
}
\details{
something
}
\examples{
PC_ODE(arg1,arg2)
}
